Try to avoid overflow!?

+++ misc/arithmexpic/genfunc.h
@@ -14,16 +14,18 @@ extern "C" {
 /*
  * Ordinary generating function for arithmetic series.
  */
-void cpx_ordinary_genfunc(cpx_t sum, cpx_t z, int prec, int (*func)(int));
+void cpx_ordinary_genfunc(cpx_t sum, cpx_t z, int prec,
+                          long (*func)(long));
 
 /*
  * Exponential generating function for arithmetic series.
  * The second form expects func to return a value in the reference
  * mpf_t*
  */
-void cpx_exponential_genfunc(cpx_t sum, cpx_t z, int prec, int (*func)(int));
+void cpx_exponential_genfunc(cpx_t sum, cpx_t z, int prec,
+                             long (*func)(long));
 void cpx_exponential_genfunc_mpf(cpx_t sum, cpx_t z, int prec,
-                                 void (*func)(mpf_t*, int));
+                                 void (*func)(mpf_t*, long));
 
 #ifdef  __cplusplus
 };
